Implement and switch to using resource sets
Many changes in here. Rather than binding shader resources (uniform
buffers, textures) directly with a separate set of binding points for
each resource type, bindings are now grouped into resource sets which
are bound together. This is in preparation for Vulkan, which uses this
model for binding resources for shaders.

We have a bunch of statically defined "standard" resource sets and
slots within them, replacing the previous uniform buffer/texture slot
definitions. We also generate resource set layouts dynamically for
Shader assets based on their parameters, and Material assets keep a
resource set based on that layout.

The GL backend dynamically creates a mapping to the GL texture/uniform
buffer binding spaces per-resource set layout. This is done such that
for the globally defined resources the binding points are the same
across all shaders that use them. This means that this is just as
efficient as the previous design in that we don't keep on rebinding the
same resources to different binding points, and where possible keep
them bound once for the entire time they are used.

As part of this we now use Vulkan-style GLSL bindings in shader source,
e.g. "layout(set = ..., binding = ...) ...". Using SPIRV-Cross, the GL
backend removes these decorations in the source that it generates and
passes to the driver, and sets up the bindings itself.

There are still a few things that can be improved here, however things
are currently working so I want to commit this as there's already
plenty enough changed here.

+++ engine/src/gpu/gl/resource.h
@@ -31,7 +31,7 @@ class GLResourceSetLayout : public GPUResourceSetLayout {
 public:
     GLResourceSetLayout(GPUResourceSetLayoutDesc &&desc);
 
-    size_t mapSlot(size_t set, size_t slot) const;
+    unsigned mapSlot(unsigned set, unsigned slot) const;
 private:
     /**
      * Mapping of slot index GL binding points.
